feat(firmware): опознание посоха, сквозные номера бросков, история во флеш
- identity.h: staffId, случайное 32-битное число в NVS, придумывается при первом
  старте. Нужен помимо MAC потому, что счётчик номеров лежит в той же NVS: если она
  потрётся, номер посоха сменится вместе со сброшенным счётчиком, и приложение
  поймёт, что нумерация началась заново. В ответ info добавлены staffId и mac.
- history.h: кольцевой файл /rolls.bin, шапка с подписью и версией формата,
  1500 записей по 36 байт. Позиция следующей записи не хранится, а вычисляется при
  старте поиском наибольшего номера. Выдача обходит кольцо со свободной ячейки,
  поэтому записи выходят по возрастанию номера без сортировки.
- Номер броска хранится в NVS и в файле сразу, при старте берётся больший: это
  переживает потерю любого одного хранилища.
- Исправлена ошибка, которая уже была в коде: sides и values в Roll были
  однобайтовыми, а .ino пропускал d до 1000, и запрос d1000 молча превращался
  в d232. Грани и значения стали двухбайтовыми, DICE_SIDES тоже.
- HISTORY_SIZE (5, история для экрана посоха) разведён с HISTORY_FLASH_SIZE (1500).
  В info поле hist теперь отдаёт ёмкость флеш-истории, раньше уходила экранная.
- Замер раздела: LittleFS 128 КБ, свободно 120 КБ. Надпись «190KB SPIFFS» в меню
  Arduino IDE неточна.
- Прошивка 0.4.0.

+++ firmware/wizards_staff/dice.h
@@ -5,15 +5,18 @@
 // Бросок кубиков. Случайные числа даёт аппаратный генератор ESP32, поэтому бросок честный.
 
 // Один бросок: сколько кубиков, каких, что выпало на каждом, сумма.
+// Грани и значения по два байта, а не по одному: протокол разрешает кубики до d1000
+// (в ТЗ, раздел 7а, это «например d3 или d1000»), а в байт влезает только 255.
+// До 2026-09-18 поля были однобайтовыми, и запрос d1000 молча превращался в d232.
 struct Roll {
-  uint8_t count;
-  uint8_t sides;
-  uint8_t values[MAX_DICE];
-  int     total;
+  uint8_t  count;
+  uint16_t sides;
+  uint16_t values[MAX_DICE];
+  int      total;
 };
 
 // Бросает count кубиков по sides граней. count ограничен MAX_DICE.
-Roll rollDice(uint8_t count, uint8_t sides) {
+Roll rollDice(uint8_t count, uint16_t sides) {
   Roll r;
   r.count = count > MAX_DICE ? MAX_DICE : count;
   r.sides = sides;
